requirements.txt
Django==3.0.7
django-ckeditor==5.9.0
django-crum==0.7.6
django-heroku==0.3.1
django-imagekit==4.0.2
django-import-export==2.2.0
djangorestframework==3.11.0
Pillow==7.1.2
Unidecode==1.1.1

# needed for heroku
# django-heroku==0.3.1

# needed for debugging purpose when debug False
django-debug-toolbar==2.2

# needed for postgreSQL
psycopg2-binary==2.8.5

# needed for deployment
gunicorn==20.0.4

# needed for automatic selenium tests
# selenium==3.141.0
